refactor: Update modal imports to new directory structure for parameterization components
